[#13449] use local repository for resolving all artifacts - fix unit tests
git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@994810 13f79535-47bb-0310-9956-ffa450edef68

+++ components/dotnet-dao-project/src/main/java/npanday/dao/impl/ProjectDaoImpl.java
@@ -747,8 +747,11 @@ public final class ProjectDaoImpl
                             artifactResolver.resolve( assembly, artifactRepositories,
                                                       localArtifactRepository );
 
-                            uacFile.getParentFile().mkdirs();
-                            FileUtils.copyFile( assembly.getFile(), uacFile );
+                            if ( assembly != null && assembly.getFile().exists() )
+                            {
+                                uacFile.getParentFile().mkdirs();
+                                FileUtils.copyFile( assembly.getFile(), uacFile );
+                            }
                         }
                         catch ( ArtifactNotFoundException e )
                         {
